fixed compile and made ecs and physics compile
