+++ src/physics/Systems.hpp
@@ -1,15 +1,13 @@
 #ifndef PHYSICS_SYSTEM
 #define PHYSICS_SYSTEM
-
-struct Transform;
-struct Rigidbody;
-class Constraint;
+#include "Components.hpp"
+#include "Constraint.hpp"
 
 #include "Register.hpp"
 
 class PhysicsSystem {
     Registry<Transform, Rigidbody> physicsRegister;
-    static PhysicsSystem* instance;
+    inline static PhysicsSystem* instance;
     std::vector<Constraint> constraints;
     PhysicsSystem();
 };
